% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ps_input.R
\name{plot.ps_input}
\alias{plot.ps_input}
\title{Plots objects of class \code{ps_input}}
\usage{
\method{plot}{ps_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ps_input}}.}

\item{...}{Further parameters (will be ignored).}
}
\value{
An object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
\code{plot.ps_input} plots a measured depth profile of
\eqn{^{210}}Pb activities along with the corresponding measurement
errors and, if available, measured \eqn{^{226}}Ra activities. The
plot may be used to assess if background \eqn{^{210}}Pb activities
are reached in order to split the data for age-depth modeling into
sections of the profile used to estimate the supported \eqn{^{210}}Pb
activity (assumed constant over time) and to estimate the age-depth
model.
}
\examples{
#

}
\references{
\insertAllCited{}
}
\seealso{
.
}
