% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_extract_samples.R
\name{ps_extract_samples}
\alias{ps_extract_samples}
\title{Creates an object of class \code{ps_input_data}}
\usage{
ps_extract_samples(x, ps_model)
}
\arguments{
\item{x}{An object of class \code{\link[rstan]{stanfit}} fitted with
\code{\link{ps_fit_model}}.}

\item{ps_model}{The corresponding \code{\link{ps_model}}
object.}
}
\value{
A \code{data.frame} object with a row for each combination
of an iter of the MCMC sampling AND each depth section (measured or
constructed artificially) and the following columns:
\describe{
  \item{iter}{An integer representing the iter of the MCMC.}
  \item{depth_profile}{A factor with two levels: \code{measured}
  if the corresponding section relates to a section of the measured
  depth profile and \code{artificial} if the corresponding section
  relates to a section of the artificially constructed depth profile
  used to model age-depth relations.}
  \item{depth_lower}{A numeric value representing the lower depth
  of the current section and MCMC iter.}
  \item{depth_upper}{A numeric value representing the upper depth
  of the current section and MCMC iter.}
  \item{age}{A numeric value representing the sampled age value for
  the current section and MCMC iter.}
  \item{depth_accumulation_rate}{A numeric value representing the sampled
  depth accumulation [yr/cm] value for the current section and MCMC
  iter. If \code{depth_profile == "measured"}, this is \code{NA} (may
  be implemented in the future).}
  \item{depth_accumulation_rate_alpha}{A numeric value representing the
  sampled depth accumulation [yr/cm] value for the current section
  and MCMC iter that is not related to the depth_accumulation of the
  previous section. If \code{depth_profile == "measured"}, this is
  \code{NA} since aggregation makes on sense here.}
  \item{pb210_tot}{A numeric value representing the sampled total
  \eqn{^{210}}Pb activity for the current section and MCMC iter.}
  \item{pb210_supported}{A numeric value representing the sampled
  supported \eqn{^{210}}Pb activity for the current section and
  MCMC iter.}
  \item{omega}{A numeric value representing the sampled
  memory parameter of the depth accumulation rates for the current
  section and MCMC iter.}
}
}
\description{
\code{ps_extract_samples} extracts samples from a fitted
\code{\link[rstan]{stanfit}} object and stored them in a
\code{data.frame}
}
\examples{
#

}
\references{
\insertAllCited{}
}
\seealso{
.
}
