% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_fit_model.R
\name{ps_fit_model}
\alias{ps_fit_model}
\title{Fits a \code{Stan} model for \eqn{^{210}}Pb based age-depth profiles.}
\usage{
ps_fit_model(ps_model, ...)
}
\arguments{
\item{ps_model}{An object of class \code{\link{ps_model}}.}

\item{...}{further parameters passed to \code{\link[rstan]{sampling}}.}
}
\value{
An object of class \code{\link[rstan]{stanfit-class}}.
}
\description{
\code{ps_fit_model} takes an object of class \code{\link{ps_model}}
and passes it to \code{\link[rstan]{sampling}} in order to do the MCMC
sampling.
}
\examples{
#

}
\references{
\insertAllCited{}
}
\seealso{
.
}
