% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_get_model.R
\name{ps_get_model}
\alias{ps_get_model}
\title{Creates a Stan model for \eqn{^{210}}Pb based age-depth profiles.}
\usage{
ps_get_model(
  data_input,
  thickness = 1,
  prior_supplyrate_alpha = 2,
  prior_supplyrate_beta = prior_supplyrate_alpha/50,
  prior_omega_alpha = 4,
  prior_omega_beta = (prior_omega_alpha - 0.4 * prior_omega_alpha)/0.4,
  prior_dar_alpha = 1.5,
  prior_dar_beta = prior_dar_alpha/15,
  prior_supported_pb210_alpha = 2,
  prior_supported_pb210_beta = prior_supported_pb210_alpha/10
)
}
\arguments{
\item{data_input}{An object of class \code{\link{ps_input}}
containing the measured \eqn{^{210}}Pb activities and optionally measured
\eqn{^{226}}Ra activities and ages inferred from \eqn{^{137}}Cs peaks from
which to construct the Bayesian age-depth model. Rows have to be ordered
in with decreasing values of \code{data_input$depth_lower}.}

\item{thickness}{A numeric value representing the thickness of the
artifial sections for which to construct the sediment accumulation
model [cm].}

\item{prior_supplyrate_alpha}{A numeric value representing the shape parameter
of the Gamma prior distribution for the atmospheric \eqn{^{210}}Pb supply rate
[Bq/kg]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_supplyrate_beta}{A numeric value representing the rate parameter
of the Gamma prior distribution for the atmospheric \eqn{^{210}}Pb supply rate
[Bq/kg]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_omega_alpha}{A numeric value representing the shape parameter
of the beta prior distribution for the depth accumulation rate memory
parameter \eqn{\omega}. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_omega_beta}{A numeric value representing the rate parameter
of the beta prior distribution for the depth accumulation rate memory
parameter \eqn{\omega}. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_dar_alpha}{A numeric value representing the shape parameter
of the Gamma prior distribution for the depth accumulation rate
[yr/cm]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_dar_beta}{A numeric value representing the rate parameter
of the Gamma prior distribution for the depth accumulation rate
[yr/cm]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_supported_pb210_alpha}{A numeric value representing the shape parameter
of the Gamma prior distribution for the auported \eqn{^{210}}Pb activity
[Bq/kg]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}

\item{prior_supported_pb210_beta}{A numeric value representing the rate parameter
of the Gamma prior distribution for the auported \eqn{^{210}}Pb activity
[Bq/kg]. The default value is chosen as reported by
\insertCite{AquinoLopez.2018;textual}{plumstan} and implemented in \code{Plum}.}
}
\value{
An object of class \code{\link{ps_model}}.
}
\description{
\code{ps_get_model} takes data of measured \eqn{^{210}}Pb activity
depth profiles and measured supported \eqn{^{210}}Pb activities as input
and constructs a Stan model that represents a Bayesian age-depth
model.
}
\details{
The Bayesian statistical framework and algorithm are derived from
\insertCite{AquinoLopez.2018;textual}{plumstan}, the corresponding R package
\href{https://github.com/maquinolopez/Plum}{\code{Plum}} and implemented
in \href{https://github.com/stan-dev/rstan}{\code{Stan} (respectively
\code{rstan})} \insertCite{Carpenter.2017, StanDevelopmentTeam.2019}{plumstan}.
}
\examples{
#

}
\references{
\insertAllCited{}
}
\seealso{
.
}
