% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_input.R
\name{ps_input}
\alias{ps_input}
\title{Creates an object of class \code{ps_input}}
\usage{
ps_input(x)
}
\arguments{
\item{x}{A \code{data.frame} with a row for each section and
data type (\eqn{^{210}}Pb or \eqn{^{226}}Ra activity or age inferred from
\eqn{^{137}}Cs data) and the following columns (in this
order):
\describe{
  \item{type}{The data type of the measurement for the current
    section. One of \code{c("pb210", "ra226" or "cs137")}.}
  \item{supported}{A logical vector indicating supported
  \eqn{^{210}}Pb or \eqn{^{226}}Ra activities (\code{TRUE}) or not
  \code{FALSE}.}
  \item{depth_lower}{The lower depth of the current section [cm].}
  \item{depth_upper}{The upper depth of the current section [cm].}
  \item{mass_density}{The mass density of the current section [g
    cm\eqn{^{-3}}].}
  \item{activity}{The measured activity [Bq kg\eqn{^{-1}}] or
  \code{NA} if no values are available.}
  \item{activity_sd}{The measurement error of the measured activity
    [Bq kg\eqn{^{-1}}] or \code{NA} if no values are available.}
  \item{age}{The age of a section or \code{NA} if such information
    is not available.}
  \item{age_sd}{The measurement error in the age estimate or \code{NA}
  if such information is not available.}
}}
}
\value{
An object of class \code{ps_input}, that is a
\code{data.frame} with a row for each section and
data type (\eqn{^{210}}Pb or \eqn{^{226}}Ra activity or age inferred from
\eqn{^{137}}Cs data) and the following columns (in this
order):
\describe{
  \item{type}{The data type of the measurement for the current
    section. One of \code{c("pb210", "ra226" or "cs137")}.}
  \item{supported}{A logical vector indicating supported
  \eqn{^{210}}Pb or \eqn{^{226}}Ra activities (\code{TRUE}) or not
  \code{FALSE}.}
  \item{depth_lower}{The lower depth of the current section [cm].}
  \item{depth_upper}{The upper depth of the current section [cm].}
  \item{mass_density}{The mass density of the current section [g
    cm\eqn{^{-3}}].}
  \item{activity}{The measured activity [Bq kg\eqn{^{-1}}] or
  \code{NA} if no values are available.}
  \item{activity_sd}{The measurement error of the measured activity
    [Bq kg\eqn{^{-1}}] or \code{NA} if no values are available.}
  \item{age}{The age of a section or \code{NA} if such information
    is not available.}
  \item{age_sd}{The measurement error in the age estimate or \code{NA}
  if such information is not available.}
}
}
\description{
\code{ps_input} is the constructor function for objects of class
\code{ps_input}. This class aims to provide a concise organisation of
various data sources needed to estimate age-depth profiles ((supported)
\eqn{^{210}}Pb activities, \eqn{^{226}}Ra activities, \eqn{^{137}}Cs activities
and ages).
}
\examples{
#

}
\references{
\insertAllCited{}
}
\seealso{
.
}
